[SERVER] refactor: 로그아웃 시 refresh token 유효성 확인 추가

+++ server/user/src/main/java/com/explorer/user/domain/user/service/impl/UserServiceImpl.java
@@ -3,6 +3,8 @@ package com.explorer.user.domain.user.service.impl;
 import com.explorer.user.domain.user.repository.UserRepository;
 import com.explorer.user.domain.user.service.UserService;
 import com.explorer.user.global.component.jwt.JwtProvider;
+import com.explorer.user.global.component.jwt.exception.JwtErrorCode;
+import com.explorer.user.global.component.jwt.exception.JwtException;
 import com.explorer.user.global.component.jwt.repository.BlackListRepository;
 import com.explorer.user.global.component.jwt.repository.RefreshTokenRepository;
 import jakarta.transaction.Transactional;
@@ -24,7 +26,15 @@ public class UserServiceImpl implements UserService {
     @Transactional
     @Override
     public void logout(Long userId, String accessToken, String refreshToken) {
-        blackListRepository.save(accessToken, jwtProvider.getExpirationDate(accessToken));
+        long remainingTime = jwtProvider.getRefreshTokenExpirationDate(refreshToken).getTime() - System.currentTimeMillis();
+        boolean isExist = refreshTokenRepository.exist(refreshToken);
+        boolean isEqual = refreshTokenRepository.find(String.valueOf(userId)).get().equals(refreshToken);
+
+        if (isExist || !isEqual || remainingTime <= 0) {
+            throw new JwtException(JwtErrorCode.INVALID_TOKEN);
+        }
+
+        blackListRepository.save(accessToken, jwtProvider.getAccessTokenExpirationDate(accessToken));
         refreshTokenRepository.delete(String.valueOf(userId));
     }
